-renamed visualizer/__main__.py to control.py and added a control class
-added new short __main__.py
